Stop tracking distribution files
It is not a good idea to version control the distribution files
as we would have to keep compiling and git adding the compiled
files. This may cause inconsistency and raise the need to git
rebase by developers.

We hop to have the distribution files compiled by a CI server
and automatically uploaded to npmjs.org